Refactor job submission to use JobManager
In the DiscordDokuWikiListener, IrcCommandListener, and DiscordMediaWikiListener classes, replaced direct job initiation with submission via the JobManager to standardize job handling across the project. The number of running jobs in the IrcCommandListener was updated to reflect the changes. The JobManager was also expanded to handle job buckets and provide lists of active and queued jobs. In the RunJob class, removed extraneous exception. This allows us to manage jobs on a more abstract level and easily fetch information about our job queues.