fix(warc): explicitly unset the Referer header
Wget-AT's default behavior seems to automatically set the Referer header to the URL of the page being called from get_urls. Since we're managing the queue in Lua instead of letting wget-at handle it, this results in weird Referer headers being sent for some pages (for example, if it happens to grab an image file and then makes an API request, the API request's referer would be set to that image file, which is weird). Setting the Referer header in the headers table to an empty string seems to resolve this behavior.